refactor: moved CLI vestingBulk
